+++ changelog.md
@@ -2,7 +2,8 @@
 
 ## 1.29.0 (TBD)
 
-- Added Entropynnium
+- Added Entropynnium and Cleansweep
+	- Cleansweep sprite thanks to [SpecialAPI](https://github.com/SpecialAPI)!
 
 ## 1.28.3 (2026-03-15)